+++ src/Sources/SourceFunction.hh
@@ -52,7 +52,7 @@ public:
    *
    * @param time_step the time at which to apply the excitation
    */
-  virtual field_t source_function(const Grid &grid, unsigned int time_step) = 0;
+  virtual field_t source_function(const Grid &grid, float time_step) = 0;
 
 };
